@@ -17,8 +17,10 @@ Gem::Specification.new do |spec|
   spec.executables   = spec.files.grep(%r{^bin/}) { |f| File.basename(f) }
   spec.test_files    = spec.files.grep(%r{^(test|spec|features)/})
   spec.require_paths = ["lib"]
-  spec.required_ruby_version = ">= 1.9.3"
+  spec.required_ruby_version = ">= 2.6.0"
 
+  spec.add_development_dependency "appraisal"
+
+  spec.add_dependency "activerecord", ">= 5.0.0"
   spec.add_dependency "wisper", "~> 2.0"
-  spec.add_dependency "activerecord", ">= 3.0.0"
 end
